requirements.txt
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-language==1.3.0
google-cloud-logging==1.14.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==2.4.4
PySocks==1.7.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tweepy==3.8.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
